[libUPnP] update Neptune & Platinum

+++ lib/libUPnP/Neptune/Source/System/Posix/NptPosixQueue.cpp
@@ -321,7 +321,7 @@ NPT_PosixQueue::Peek(NPT_QueueItem*& item, NPT_Timeout timeout)
 NPT_GenericQueue*
 NPT_GenericQueue::CreateInstance(NPT_Cardinal max_items)
 {
-    NPT_LOG_FINER_1("queue max_items = %ld", max_items);
+    NPT_LOG_FINER_1("queue max_items = %d", (int)max_items);
     return new NPT_PosixQueue(max_items);
 }
